add a "start" script in package.json
In preparation for PaaSing our little app, the npm needs to be able
to start the software on its standard way.

This will enable us later on to push automatical deployement using
git hooks and push notifications or a polling agent at the
Raspberry Pi end. Perhaps we could run the app on Heroku even though
it makes very little sense.